+++ pkg/api/v1/clients/client_interface.go
@@ -178,20 +178,3 @@ func (o WatchOpts) WithDefaults() WatchOpts {
 	}
 	return o
 }
-
-func TranslateWatchOptsIntoListOpts(wopts WatchOpts) ListOpts {
-	clopts := ListOpts{Ctx: wopts.Ctx, ExpressionSelector: wopts.ExpressionSelector, Selector: wopts.Selector}
-	return clopts
-}
-
-// TranslateResourceNamespaceListToListOptions translates the resource namespace list options to List Options
-func TranslateResourceNamespaceListToListOptions(lopts resources.ResourceNamespaceListOptions) ListOpts {
-	clopts := ListOpts{Ctx: lopts.Ctx, ExpressionSelector: lopts.ExpressionSelector}
-	return clopts
-}
-
-// TranslateResourceNamespaceListToWatchOptions translates the resource namespace watch options to Watch Options
-func TranslateResourceNamespaceListToWatchOptions(wopts resources.ResourceNamespaceWatchOptions) WatchOpts {
-	clopts := WatchOpts{Ctx: wopts.Ctx, ExpressionSelector: wopts.ExpressionSelector}
-	return clopts
-}
